Grids: rename m_state_storing_core
