+++ packages/devextreme/js/__internal/grids/pivot_grid/data_controller/m_data_controller.ts
@@ -5,7 +5,7 @@ import { extend } from '@js/core/utils/extend';
 import { each, map } from '@js/core/utils/iterator';
 import { format } from '@js/core/utils/string';
 import { isDefined, isString } from '@js/core/utils/type';
-import stateStoring from '@ts/grids/grid_core/state_storing/m_state_storing_core';
+import stateStoring from '@ts/grids/grid_core/state_storing/m_state_storing_controller';
 import { createColumnsInfo, foreachColumnInfo } from '@ts/grids/grid_core/virtual_columns/m_virtual_columns_core';
 import VirtualScrollControllerModule from '@ts/grids/grid_core/virtual_scrolling/m_virtual_scrolling_core';
